Value range refinement using uni-variate expressions.
This PR introduces value range refinement of shape symbols by symbolically evaluating the
value range of the involved guards. This should help `_maybe_evaluate_static` to eliminate
more guards.

This is a stack of PRs created from the discussion on: #96616.

In summary, this PR:
- simplifies `FloorDiv` nodes on the left-hand side of an expression so as to isolate a
symbol in the numerator
- tries to match the expression against the form: `<symbol> <relop> <expr>`
- uses the matched expression for refining the value range of `<symbol>` using the range
of `<expr>`

ghstack-source-id: 3c2a79230022b6a4d5d69f70c28129cfbb56e54f
Pull Request resolved: https://github.com/pytorch/pytorch/pull/97963