Improve repeat_interleave with scalar repeat value
`repeat_interleave_symint` is currently implemented by guarding on the `SymInt`
and converting it to a tensor to pass to the Tensor overload. This instead
implements it as a copy of an expanded tensor, which can be done without guards
and is also much more efficient in eager mode to boot.

ghstack-source-id: b9ce7b0d12ae7fe3cbd9c60e786b7d4c370d2152
Pull Request resolved: https://github.com/pytorch/pytorch/pull/102570