Make TestPythonRegistration clean up after itself
We did this for TestCustomOp, now we are applying the same thing to
TestPythonRegistration.

This PR:
- changes TestPythonRegistration to register new ops under a single
namespace (self.test_ns)
- clean up the namespace by deleting it from torch.ops after each test
is done running.

This avoids a problem where if an op is re-defined, torch.ops.myns.op
crashes because we do some caching. The workaround in many of these
tests have been to just create an op with a different name, but this PR
makes it so that we don't need to do this.

Test Plan:
- existing tests

ghstack-source-id: d00f919d3858a038f05b1efc3fa8116520c95c14
Pull Request resolved: https://github.com/pytorch/pytorch/pull/102292